+++ CHANGELOG.md
@@ -7,6 +7,7 @@ a concise history of changes made.
 
 ## Latest
 - Added an About page, describing the site.
+- Added a field to the Species page, so you can search for species, subspecies, etc.
 - Fixed error where the species total was also counting subspecies, forms, etc.
 
 ## 2025.06.10